+++ src/edu/dottn/services/ServicePost.java
@@ -17,6 +17,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 
 /**
  *
@@ -27,28 +29,89 @@ public class ServicePost implements PService<Post> {
     Connection cnx = MyConnection.getInstance().getConnection();
 
     public void ajouter(Post p) {
+        // Vérifier que les champs ne sont pas nuls
+        if (p.getTitlePost().isEmpty()) {
+            Alert alert = new Alert(AlertType.WARNING);
+            alert.setTitle("Missing field");
+            alert.setHeaderText(null);
+            alert.setContentText("The title field must not be empty");
+            alert.showAndWait();
+            return;
+        }
+        if (p.getDescription().isEmpty()) {
+            Alert alert = new Alert(AlertType.WARNING);
+            alert.setTitle("Empty field");
+            alert.setHeaderText(null);
+            alert.setContentText("The description field must not be empty");
+            alert.showAndWait();
+            return;
+        }
+
+        // Vérifier que les champs respectent une longueur maximale
+        int titleMaxLen = 50;
+        int descMaxLen = 500;
+        if (p.getTitlePost().length() > titleMaxLen) {
+            Alert alert = new Alert(AlertType.WARNING);
+            alert.setTitle("field too long");
+            alert.setHeaderText(null);
+            alert.setContentText("The title field must not exceed " + titleMaxLen + " characters");
+            alert.showAndWait();
+            return;
+            
+        }
+        if (p.getDescription().length() > descMaxLen) {
+            Alert alert = new Alert(AlertType.WARNING);
+            alert.setTitle("field too long");
+            alert.setHeaderText(null);
+            alert.setContentText("The description field must not exceed  " + descMaxLen + " characters");
+            alert.showAndWait();
+            return;
+        }
+
         String sql = "INSERT INTO post (titlePost, description) VALUES (?, ?)";
         try (PreparedStatement statement = cnx.prepareStatement(sql)) {
             statement.setString(1, p.getTitlePost());
             statement.setString(2, p.getDescription());
             int rowsAffected = statement.executeUpdate();
             if (rowsAffected > 0) {
-                System.out.println("Post added successfully");
+                Alert alert = new Alert(AlertType.INFORMATION);
+                alert.setTitle("Post added ");
+                alert.setHeaderText(null);
+                alert.setContentText("The post was successfully added");
+                alert.showAndWait();
             } else {
-                System.out.println("Post add failed");
+                Alert alert = new Alert(AlertType.ERROR);
+                alert.setTitle("Erreur");
+                alert.setHeaderText(null);
+                alert.setContentText("L'ajout du post a échoué");
+                alert.showAndWait();
             }
         } catch (SQLException ex) {
-            System.err.println("Error while adding post: " + ex.getMessage());
+            Alert alert = new Alert(AlertType.ERROR);
+            alert.setTitle("Error");
+            alert.setHeaderText(null);
+            alert.setContentText("Error adding post:" + ex.getMessage());
+            alert.showAndWait();
         }
     }
 
     public void supprimerParId(int idPost) {
+        // Vérifier que l'ID de l'article est valide
+        if (idPost <= 0) {
+            Alert alert = new Alert(AlertType.WARNING);
+            alert.setTitle("invalid ID");
+            alert.setHeaderText(null);
+            alert.setContentText("Item ID must be a non-zero positive integer");
+            alert.showAndWait();
+            return;
+        }
+
         try {
             String req = "DELETE FROM `post` WHERE idPost = " + idPost;
             Statement st = cnx.createStatement();
             int rowsAffected = st.executeUpdate(req);
             if (rowsAffected > 0) {
-                System.out.println("Post deleted");
+                System.out.println("Post Deleted !!");
             } else {
                 System.out.println("Post delete failed");
             }
@@ -58,6 +121,24 @@ public class ServicePost implements PService<Post> {
     }
 
     public void modifier(Post p) {
+        // Vérifier que les champs ne sont pas nuls
+        if (p.getTitlePost().isEmpty()) {
+            Alert alert = new Alert(AlertType.WARNING);
+            alert.setTitle("Champ titre invalide");
+            alert.setHeaderText(null);
+            alert.setContentText("The title field must not be empty");
+            alert.showAndWait();
+            return;
+        }
+        if (p.getDescription().isEmpty()) {
+            Alert alert = new Alert(AlertType.WARNING);
+            alert.setTitle("Champ description invalide");
+            alert.setHeaderText(null);
+            alert.setContentText("The Description field must not be empty");
+            alert.showAndWait();
+            return;
+        }
+
         try {
             String req = "UPDATE `post` SET `titlePost` = ?, `description` = ? WHERE idPost=?";
             PreparedStatement statement = cnx.prepareStatement(req);
@@ -66,9 +147,9 @@ public class ServicePost implements PService<Post> {
             statement.setString(2, p.getDescription());
             statement.setInt(3, p.getIdPost());
             statement.executeUpdate();
-            System.out.println("Post MODIFIÉ !");
+            System.out.println("Post edited!");
         } catch (SQLException ex) {
-            System.out.println("Erreur lors de la modification du post: " + ex.getMessage());
+            System.out.println("Error editing post:" + ex.getMessage());
         }
     }
 
@@ -83,9 +164,10 @@ public class ServicePost implements PService<Post> {
                 list.add(p);
             }
         } catch (SQLException ex) {
-            System.out.println("Erreur lors de la récupération de tous les posts: " + ex.getMessage());
+            System.out.println("Error retrieving all posts: " + ex.getMessage());
         }
 
+        
         return list;
     }
 
@@ -100,19 +182,23 @@ public class ServicePost implements PService<Post> {
                 post = new Post(rs.getString("titlePost"), rs.getString("description"), rs.getTimestamp("date_created"));
             }
         } catch (SQLException ex) {
-            System.out.println("Erreur lors de la récupération du post avec l'id " + id + ": " + ex.getMessage());
+            System.out.println("Error retrieving post with l'id: " + id + ": " + ex.getMessage());
         }
 
         return post;
     }
 
     public Post getRecentPost() {
+       
         Post recentPost = null;
+
         try {
+            
             String req = "SELECT * FROM post ORDER BY date_created DESC LIMIT 1";
             Statement st = cnx.createStatement();
             ResultSet rs = st.executeQuery(req);
 
+            // Extraction des résultats de la requête SQL et conversion en une liste de Post
             List<Post> posts = Stream.of(rs)
                     .flatMap(resultSet -> {
                         try {
@@ -131,35 +217,50 @@ public class ServicePost implements PService<Post> {
                     })
                     .collect(Collectors.toList());
 
+            // Si la liste de Post extraite n'est pas vide, le premier élément est le post le plus récent
             if (!posts.isEmpty()) {
                 recentPost = posts.get(0);
+            } else {
+                // Si aucun post n'est trouvé, afficher une alerte
+                Alert alert = new Alert(AlertType.INFORMATION);
+                alert.setTitle("No Posts Found");
+                alert.setHeaderText(null);
+                alert.setContentText("No posts were found in the database.");
+                alert.showAndWait();
             }
         } catch (SQLException ex) {
-            System.out.println("Erreur " + ex.getMessage());
+            // Si une erreur se produit, afficher une alerte avec le message d'erreur
+            Alert alert = new Alert(AlertType.ERROR);
+            alert.setTitle("Error");
+            alert.setHeaderText(null);
+            alert.setContentText("An error occurred while fetching the most recent post: " + ex.getMessage());
+            alert.showAndWait();
         }
+        // Retourner le post le plus récent ou null si aucun post n'est trouvé
         return recentPost;
-    }
+}
+
 
     public List<Post> searchPostsByKeyword(String keyword) {
-        List<Post> posts = new ArrayList<>();
-        List<Post> result = new ArrayList<>();
-        try {
-            String req = "SELECT * FROM post";
-            Statement st = cnx.createStatement();
-            ResultSet rs = st.executeQuery(req);
-            while (rs.next()) {
-                posts.add(new Post(rs.getString("titlePost"), rs.getString("description"), rs.getTimestamp("date_created")));
-            }
-        } catch (SQLException ex) {
-            System.out.println("Erreur" + ex.getMessage());
+    List<Post> posts = new ArrayList<>();  // Créer une liste vide pour stocker les objets Post
+    List<Post> result = new ArrayList<>();  // Créer une autre liste vide pour stocker les résultats de la recherche
+    try {
+        String req = "SELECT * FROM post";  
+        Statement st = cnx.createStatement();  
+        ResultSet rs = st.executeQuery(req);  
+        while (rs.next()) {  
+            posts.add(new Post(rs.getString("titlePost"), rs.getString("description"), rs.getTimestamp("date_created")));
         }
-
-        result = posts.stream()
-                .filter(p -> p.getTitlePost().contains(keyword) || p.getDescription().contains(keyword))
-                .collect(Collectors.toList());
-
-        return result;
+    } catch (SQLException ex) {
+        System.out.println("Erreur" + ex.getMessage());  // 
     }
+
+    result = posts.stream()  // Utiliser Stream pour traiter les objets Post de la liste "posts"
+            .filter(p -> p.getTitlePost().contains(keyword) || p.getDescription().contains(keyword))  // Filtrer la liste pour ne garder que les objets Post contenant le mot-clé
+            .collect(Collectors.toList());  // Stocker les objets filtrés dans la liste "result"
+
+    return result;  
+}
 
     public List<Post> getAllPostsSortedByDate(boolean isAscending) {
         List<Post> posts = new ArrayList<>();
@@ -173,7 +274,7 @@ public class ServicePost implements PService<Post> {
                 Timestamp dateCreated = rs.getTimestamp("date_created");
                 Post post = new Post(title, description, dateCreated);
                 posts.add(post);
-                
+
             }
         } catch (SQLException ex) {
             System.out.println("Erreur " + ex.getMessage());
